@@ -5,7 +5,7 @@ import org.example.ai.chatbot.domain.rebate.port.IRebatePort;
 
 import org.springframework.stereotype.Service;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 
 
 
