@@ -1,12 +1,12 @@
 package org.example.ai.chatbot.domain.openai.service;
 
 import org.example.ai.chatbot.domain.openai.model.aggregates.ChatProcessAggregate;
-import org.example.ai.chatbot.domain.openai.model.aggregates.ChatProcessAggregate;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyEmitter;
+import reactor.core.publisher.Flux;
 
 
 public interface IChatService {
 
-    ResponseBodyEmitter completions(ResponseBodyEmitter emitter, ChatProcessAggregate chatProcess);
+    Flux<String> generateStream(ChatProcessAggregate chatProcessAggregate);
 
 }
